PR Review, better node removal
