+++ src/test/java/org/sonarlint/intellij/ui/tree/SecurityHotspotTreeModelBuilderTests.java
@@ -21,7 +21,6 @@ package org.sonarlint.intellij.ui.tree;
 
 import com.intellij.openapi.editor.Document;
 import com.intellij.openapi.editor.RangeMarker;
-import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiFile;
 import java.util.ArrayList;
@@ -36,8 +35,10 @@ import javax.annotation.Nullable;
 import javax.swing.tree.DefaultTreeModel;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.sonarlint.intellij.AbstractSonarLintLightTests;
 import org.sonarlint.intellij.actions.filters.FilterSecurityHotspotSettings;
 import org.sonarlint.intellij.actions.filters.SecurityHotspotFilters;
+import org.sonarlint.intellij.editor.CodeAnalyzerRestarter;
 import org.sonarlint.intellij.finding.hotspot.LiveSecurityHotspot;
 import org.sonarlint.intellij.ui.nodes.AbstractNode;
 import org.sonarlint.intellij.ui.nodes.LiveSecurityHotspotNode;
@@ -51,11 +52,11 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-class SecurityHotspotTreeModelBuilderTests {
+class SecurityHotspotTreeModelBuilderTests extends AbstractSonarLintLightTests {
 
+  private final CodeAnalyzerRestarter codeAnalyzerRestarter = mock(CodeAnalyzerRestarter.class);
   private SecurityHotspotTreeModelBuilder treeBuilder;
   private DefaultTreeModel model;
-  private Project project = mock(Project.class);
 
   @BeforeEach
   void init() {
@@ -63,6 +64,7 @@ class SecurityHotspotTreeModelBuilderTests {
     model = treeBuilder.createModel();
     FilterSecurityHotspotSettings.setCurrentlySelectedFilter(SecurityHotspotFilters.DEFAULT_FILTER);
     FilterSecurityHotspotSettings.setResolved(false);
+    replaceProjectService(CodeAnalyzerRestarter.class, codeAnalyzerRestarter);
   }
 
   @Test
@@ -141,13 +143,13 @@ class SecurityHotspotTreeModelBuilderTests {
 
     treeBuilder.updateModelWithoutFileNode(data, "empty");
 
-    assertThat(treeBuilder.filterSecurityHotspots(project, SecurityHotspotFilters.SHOW_ALL)).isEqualTo(3);
+    assertThat(treeBuilder.filterSecurityHotspots(getProject(), SecurityHotspotFilters.SHOW_ALL)).isEqualTo(3);
     assertThat(treeBuilder.getFilteredNodes()).hasSize(3);
 
-    assertThat(treeBuilder.filterSecurityHotspots(project, SecurityHotspotFilters.EXISTING_ON_SONARQUBE)).isEqualTo(2);
+    assertThat(treeBuilder.filterSecurityHotspots(getProject(), SecurityHotspotFilters.EXISTING_ON_SONARQUBE)).isEqualTo(2);
     assertThat(treeBuilder.getFilteredNodes()).hasSize(2);
 
-    assertThat(treeBuilder.filterSecurityHotspots(project, SecurityHotspotFilters.LOCAL_ONLY)).isEqualTo(1);
+    assertThat(treeBuilder.filterSecurityHotspots(getProject(), SecurityHotspotFilters.LOCAL_ONLY)).isEqualTo(1);
     assertThat(treeBuilder.getFilteredNodes()).hasSize(1);
   }
 
@@ -163,11 +165,11 @@ class SecurityHotspotTreeModelBuilderTests {
     treeBuilder.updateModelWithoutFileNode(data, "empty");
 
     FilterSecurityHotspotSettings.setResolved(false);
-    assertThat(treeBuilder.filterSecurityHotspots(project, SecurityHotspotFilters.SHOW_ALL)).isEqualTo(2);
+    assertThat(treeBuilder.filterSecurityHotspots(getProject(), SecurityHotspotFilters.SHOW_ALL)).isEqualTo(2);
     assertThat(treeBuilder.getFilteredNodes()).hasSize(2);
 
     FilterSecurityHotspotSettings.setResolved(true);
-    assertThat(treeBuilder.filterSecurityHotspots(project, SecurityHotspotFilters.SHOW_ALL)).isEqualTo(4);
+    assertThat(treeBuilder.filterSecurityHotspots(getProject(), SecurityHotspotFilters.SHOW_ALL)).isEqualTo(4);
     assertThat(treeBuilder.getFilteredNodes()).hasSize(4);
   }
